+++ apps/member/src/components/manage/ManageLibrarySection/ManageLibrarySection.tsx
@@ -6,7 +6,7 @@ import ActionButton from '@components/common/ActionButton/ActionButton';
 import Pagination from '@components/common/Pagination/Pagination';
 import { Section } from '@components/common/Section';
 import BookLoanConditionStatusBadge from '@components/library/BookLoanConditionStatusBadge/BookLoanConditionStatusBadge';
-import MemberInfoModal from '@components/modal/MemberInfoModal/MemberInfoModal';
+import { MemberInfoModal } from '@components/modal';
 
 import { TABLE_HEAD } from '@constants/head';
 import useModal from '@hooks/common/useModal';
